Reorganize client files, fix pixi unload on quit
