+++ src/client/js/lib/cookies.js
